Refactored for pulling WiW accounts from up to 21 days back and using email only to ID users across Canvas and WiW.
